Add mysql sizes and name according to convention
